requirements.txt
kafka-python==2.0.2
google-generativeai==0.3.2
python-dotenv==1.0.0
Faker==20.1.0
colorama==0.4.6
